requirements.txt
websockets==10.4
paho-mqtt==1.6.1
pyserial==3.5